Move to new background template sending
By defining the templates as HTML I can more easily convey the natural
flow of the components, then the actual component that I create with it
loads the template and attaches things to the semantically meaningful
classes/ids that are present.

This effectively separates the presentational and behavioral logic of
the application. I can still add classes to do things like transitions
cleanly from the JS, but they can be defined alongside the html easily.